Working file watcher!
Checks for modified (timestamp), created (new entries), and deleted (not updated during sweep) files. The latter requires some... interesting use of data structures to avoid becoming n^2 (the naive approach would be to just push two lists of files, but diffing them (set difference) would probably be fairly inefficient and/or complex. With that said the current impl is not fantastic (modeling the file tree(s) with custom structs would be much better), but for now it works.